+++ Seashore/source/application/SeaApplication.m
@@ -2,7 +2,7 @@
 
 @implementation SeaApplication
 
-- (unsigned int)validModesForFontPanel:(NSFontPanel *)fontPanel
+- (NSUInteger)validModesForFontPanel:(NSFontPanel *)fontPanel
 {
 	return NSFontPanelFaceModeMask | NSFontPanelSizeModeMask | NSFontPanelCollectionModeMask;
 }
